LayoutTests/editing/pasteboard/copy-paste-with-shadow-content-expected.txt

This tests copying and pasting content with a shadow tree.
To test manually, copy paste the content in the blue box to the green box below. All the text shoul be copied & pasted

pasted:
| <span>
|   id="host"
|   "hello "
|   <span>
|     id="host"
|     <span>
|       style="display: contents;"
|       "world"
|     " "
|     "WebKit"
|   " rocks<#selection-caret>"

text/html:
| <span>
|   id="host"
|   "hello "
|   <span>
|     id="host"
|     <span>
|       style="display: contents;"
|       "world"
|     " "
|     "WebKit"
|   " rocks<#selection-caret>"

text/plain:
| "hello world WebKit rocks"
